refactor: `parser` `parser_v2` > `parser::{v1, v2}`
