+++ clarity/src/vm/ast/definition_sorter/tests.rs
@@ -38,7 +38,7 @@ fn test_clarity_versions_definition_sorter(#[case] version: ClarityVersion) {}
 
 fn run_scoped_parsing_helper(contract: &str, version: ClarityVersion) -> ParseResult<ContractAST> {
     let contract_identifier = QualifiedContractIdentifier::transient();
-    let pre_expressions = parser::parse(contract)?;
+    let pre_expressions = parser::v1::parse(contract)?;
     let mut contract_ast = ContractAST::new(contract_identifier.clone(), pre_expressions);
     ExpressionIdentifier::run_pre_expression_pass(&mut contract_ast, version)?;
     DefinitionSorter::run_pass(&mut contract_ast, &mut (), version)?;
